api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
asyncpg==0.29.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
orjson==3.10.7
starlette==0.38.2